Added Refactored DropIn with session flow for demo app
